Extracts Printout code from Generator to new file
